transforme in ts
